rename module to skill_grid
